Relocate functional test app and revise code formatting
