+++ Tests/Functional/BaseTestCase.php
@@ -6,8 +6,19 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class BaseTestCase extends WebTestCase
 {
+    protected static function getKernelClass()
+    {
+        require_once __DIR__.'/app/AppKernel.php';
+
+        return 'Jmikola\AutoLoginBundle\Tests\Functional\app\AppKernel';
+    }
+
     protected static function createKernel(array $options = array())
     {
-        return new AppKernel(isset($options['config']) ? $options['config'] : 'default.yml');
+        $class = self::getKernelClass();
+
+        return new $class(
+            isset($options['config']) ? $options['config'] : 'default.yml'
+        );
     }
 }
