+++ Tests/Functional/BundleInitializationTest.php
@@ -4,10 +4,7 @@ namespace Jmikola\AutoLoginBundle\Tests\Functional;
 
 class BundleInitializationTest extends BaseTestCase
 {
-    /**
-     * @test
-     */
-    public function bundle_will_install_with_no_errors()
+    public function testRegisterBundle()
     {
         static::createClient();
     }
